+++ examples/print/src/print_rs/lib.rs
@@ -1,4 +1,0 @@
-#[ic_cdk::query]
-fn print() {
-    ic_cdk::print("Hello World");
-}