+++ library/ic-ledger-types/CHANGELOG.md
@@ -6,6 +6,70 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.15.0] - 2025-04-22
+
+### Changed
+
+- Upgrade `ic-cdk` to v0.18.
+
+## [0.14.0] - 2024-11-08
+
+### Changed
+
+- Upgrade `ic-cdk` to v0.17.
+
+### Added
+- as_bytes method to AccountIdentifier in ic-ledger-types
+
+## [0.13.0] - 2024-08-27
+
+### Changed
+
+- Upgrade `ic-cdk` to v0.16.
+
+## [0.12.0] - 2024-07-01
+
+### Changed
+
+- Upgrade `ic-cdk` to v0.15.
+
+## [0.11.0] - 2024-05-17
+
+### Changed
+
+- Upgrade `ic-cdk` to v0.14.
+
+## [0.10.0] - 2024-03-01
+
+### Changed
+- Upgrade `ic-cdk` to v0.13.
+
+## [0.9.0] - 2023-11-23
+
+### Changed
+- Upgrade `ic-cdk` to v0.12 and `candid` to v0.10.
+
+## [0.8.0] - 2023-09-18
+
+### Changed
+- Upgrade `ic-cdk` to v0.11.
+
+## [0.7.0] - 2023-07-13
+
+### Added
+- from_hex/from_slice/to_hex methods to AccountIdentifier in ic-ledger-types
+
+### Changed
+- Upgrade `ic-cdk` to v0.10 and `candid` to v0.9.
+
+## [0.6.0] - 2023-06-20
+### Changed
+- Upgrade `ic-cdk` to v0.9.
+
+## [0.5.0] - 2023-05-26
+### Changed
+- Upgrade `ic-cdk` to v0.8.
+
 ## [0.4.2] - 2023-03-01
 ### Fixed
 - Fill missing docs.